数值类型修改

+++ YYKit/Text/Component/YYTextLayout.m
@@ -535,11 +535,11 @@ dispatch_semaphore_signal(_lock);
         
         if (constraintSizeIsExtended) {
             if (isVerticalForm) {
-                if (rect.origin.x + rect.size.width >
+                if ((int)(rect.origin.x + rect.size.width) >
                     constraintRectBeforeExtended.origin.x +
                     constraintRectBeforeExtended.size.width) break;
             } else {
-                if (rect.origin.y + rect.size.height >
+                if ((int)(rect.origin.y + rect.size.height) >
                     constraintRectBeforeExtended.origin.y +
                     constraintRectBeforeExtended.size.height) break;
             }
